@@ -6,29 +6,29 @@ class TextStyleHtml {
   final Iterable<dynamic> _deps;
 
   /// The line height.
-  final double height;
+  final double? height;
 
   /// The number of max lines that should be rendered.
-  final int maxLines;
+  final int? maxLines;
 
   /// The parent style.
-  final TextStyleHtml parent;
+  final TextStyleHtml? parent;
 
   /// The input [TextStyle].
-  final TextStyle style;
+  final TextStyle? style;
 
   /// The text alignment.
-  final TextAlign textAlign;
+  final TextAlign? textAlign;
 
   /// The text direction.
-  final TextDirection textDirection;
+  final TextDirection? textDirection;
 
   /// The overflow behavior.
-  final TextOverflow textOverflow;
+  final TextOverflow? textOverflow;
 
   /// Creates a text style.
   TextStyleHtml({
-    @required Iterable deps,
+    required Iterable deps,
     this.height,
     this.maxLines,
     this.parent,
@@ -40,18 +40,18 @@ class TextStyleHtml {
 
   /// Creates the root text style.
   factory TextStyleHtml.root(
-      Iterable<dynamic> deps, TextStyle widgetTextStyle) {
+      Iterable<dynamic> deps, TextStyle? widgetTextStyle) {
     var style = _getDependency<TextStyle>(deps);
     if (widgetTextStyle != null) {
       style = widgetTextStyle.inherit
-          ? style.merge(widgetTextStyle)
+          ? style!.merge(widgetTextStyle)
           : widgetTextStyle;
     }
 
-    var mqd = _getDependency<MediaQueryData>(deps);
+    var mqd = _getDependency<MediaQueryData>(deps)!;
     final tsf = mqd.textScaleFactor;
     if (tsf != 1) {
-      style = style.copyWith(fontSize: style.fontSize * tsf);
+      style = style!.copyWith(fontSize: style.fontSize! * tsf);
     }
 
     return TextStyleHtml(
@@ -67,18 +67,18 @@ class TextStyleHtml {
   /// This needs to be done because
   /// `TextStyle` with existing height cannot be copied with `height=null`.
   /// See [flutter/flutter#58765](https://github.com/flutter/flutter/issues/58765).
-  TextStyle get styleWithHeight =>
-      height != null && height >= 0 ? style.copyWith(height: height) : style;
+  TextStyle? get styleWithHeight =>
+      height != null && height! >= 0 ? style!.copyWith(height: height) : style;
 
   /// Creates a copy with the given fields replaced with the new values.
   TextStyleHtml copyWith({
-    double height,
-    int maxLines,
-    TextStyleHtml parent,
-    TextStyle style,
-    TextAlign textAlign,
-    TextDirection textDirection,
-    TextOverflow textOverflow,
+    double? height,
+    int? maxLines,
+    TextStyleHtml? parent,
+    TextStyle? style,
+    TextAlign? textAlign,
+    TextDirection? textDirection,
+    TextOverflow? textOverflow,
   }) =>
       TextStyleHtml(
         deps: _deps,
@@ -94,9 +94,9 @@ class TextStyleHtml {
   /// Gets dependency value by type.
   ///
   /// See [WidgetFactory.getDependencies].
-  T getDependency<T>() => _getDependency<T>(_deps);
+  T? getDependency<T>() => _getDependency<T>(_deps);
 
-  static T _getDependency<T>(Iterable<dynamic> deps) {
+  static T? _getDependency<T>(Iterable<dynamic> deps) {
     for (final value in deps.whereType<T>()) {
       return value;
     }
@@ -108,33 +108,33 @@ class TextStyleHtml {
 /// A text styling builder.
 class TextStyleBuilder<T1> {
   /// The parent builder.
-  final TextStyleBuilder parent;
+  final TextStyleBuilder? parent;
 
-  List<Function> _builders;
-  List _inputs;
-  TextStyleHtml _parentOutput;
-  TextStyleHtml _output;
+  List<Function>? _builders;
+  List? _inputs;
+  TextStyleHtml? _parentOutput;
+  TextStyleHtml? _output;
 
   /// Create a builder.
   TextStyleBuilder({this.parent});
 
   /// Enqueues a callback.
   void enqueue<T2>(
-    TextStyleHtml Function(TextStyleHtml tsh, T2 input) builder, [
-    T2 input,
+    TextStyleHtml? Function(TextStyleHtml? tsh, T2 input)? builder, [
+    T2? input,
   ]) {
     if (builder == null) return;
 
     assert(_output == null, 'Cannot add builder after being built');
     _builders ??= [];
-    _builders.add(builder);
+    _builders!.add(builder);
 
     _inputs ??= [];
-    _inputs.add(input);
+    _inputs!.add(input);
   }
 
   /// Builds a [TextStyleHtml] by calling queued callbacks.
-  TextStyleHtml build(BuildContext context) {
+  TextStyleHtml? build(BuildContext context) {
     final parentOutput = parent?.build(context);
     if (parentOutput == null || parentOutput != _parentOutput) {
       _parentOutput = parentOutput;
@@ -145,10 +145,10 @@ class TextStyleBuilder<T1> {
     if (_builders == null) return _output = _parentOutput;
 
     _output = _parentOutput?.copyWith(parent: _parentOutput);
-    final l = _builders.length;
+    final l = _builders!.length;
     for (var i = 0; i < l; i++) {
-      final builder = _builders[i];
-      _output = builder(_output, _inputs[i]);
+      final builder = _builders![i];
+      _output = builder(_output, _inputs![i]);
       assert(_output?.parent == _parentOutput);
     }
 
@@ -156,19 +156,19 @@ class TextStyleBuilder<T1> {
   }
 
   /// Returns `true` if this shares same styling with [other].
-  bool hasSameStyleWith(TextStyleBuilder other) {
+  bool hasSameStyleWith(TextStyleBuilder? other) {
     if (other == null) return false;
 
-    var thisWithBuilder = this;
+    TextStyleBuilder thisWithBuilder = this;
     while (thisWithBuilder._builders == null) {
       if (thisWithBuilder.parent == null) break;
-      thisWithBuilder = thisWithBuilder.parent;
+      thisWithBuilder = thisWithBuilder.parent!;
     }
 
     var otherWithBuilder = other;
     while (otherWithBuilder._builders == null) {
       if (otherWithBuilder.parent == null) break;
-      otherWithBuilder = otherWithBuilder.parent;
+      otherWithBuilder = otherWithBuilder.parent!;
     }
 
     return thisWithBuilder == otherWithBuilder;
