@@ -3,23 +3,23 @@ import 'package:flutter/widgets.dart';
 import '../core_data.dart';
 import '../core_helpers.dart';
 
-class HeightPlaceholder extends WidgetPlaceholder<CssLength> {
-  final TextStyleBuilder tsb;
+class HeightPlaceholder extends WidgetPlaceholder<CssLength?> {
+  final TextStyleBuilder? tsb;
 
-  final List<CssLength> _heights = [];
+  final List<CssLength?> _heights = [];
 
-  HeightPlaceholder(CssLength height, this.tsb) : super(height) {
-    super.wrapWith((c, w) => _build(c, w, height, tsb));
+  HeightPlaceholder(CssLength? height, this.tsb) : super(height) {
+    super.wrapWith((c, w) => _build(c, w, height!, tsb!));
     _heights.add(height);
   }
 
-  CssLength get height => _heights.first;
+  CssLength? get height => _heights.first;
 
   void mergeWith(HeightPlaceholder other) {
     final height = other.height;
     _heights.add(height);
 
-    super.wrapWith((c, w) => _build(c, w, height, other.tsb));
+    super.wrapWith((c, w) => _build(c, w, height!, other.tsb!));
   }
 
   @override
@@ -30,7 +30,7 @@ class HeightPlaceholder extends WidgetPlaceholder<CssLength> {
       TextStyleBuilder tsb) {
     final existing = child is SizedBox ? child.height : 0.0;
     final value = height.getValue(tsb.build(context));
-    if (value != null && value > existing) return SizedBox(height: value);
+    if (value != null && value > existing!) return SizedBox(height: value);
     return child;
   }
 }
