@@ -6,9 +6,8 @@ import 'package:flutter/widgets.dart';
 /// A CSS block.
 class CssBlock extends CssSizing {
   /// Creates a CSS block.
-  CssBlock({@required Widget child, Key key})
-      : assert(child != null),
-        super(child: child, key: key);
+  CssBlock({required Widget child, Key? key})
+      :        super(child: child, key: key);
 
   CssSizingValue get _100percent => const _CssSizingPercentage(100);
 
@@ -24,16 +23,16 @@ class CssBlock extends CssSizing {
 /// A CSS sizing widget.
 class CssSizing extends SingleChildRenderObjectWidget {
   /// The maximum height.
-  final CssSizingValue maxHeight;
+  final CssSizingValue? maxHeight;
 
   /// The maximum width.
-  final CssSizingValue maxWidth;
+  final CssSizingValue? maxWidth;
 
   /// The minimum height.
-  final CssSizingValue minHeight;
+  final CssSizingValue? minHeight;
 
   // The minimum width;
-  final CssSizingValue minWidth;
+  final CssSizingValue? minWidth;
 
   /// The preferred axis.
   ///
@@ -43,18 +42,18 @@ class CssSizing extends SingleChildRenderObjectWidget {
   /// the [preferredHeight] will be used for sizing.
   ///
   /// By default (`null` preferred axis), [preferredWidth] will be used.
-  final Axis preferredAxis;
+  final Axis? preferredAxis;
 
   /// The preferred height.
-  final CssSizingValue preferredHeight;
+  final CssSizingValue? preferredHeight;
 
   /// The preferred width.
-  final CssSizingValue preferredWidth;
+  final CssSizingValue? preferredWidth;
 
   /// Creates a CSS sizing.
   CssSizing({
-    @required Widget child,
-    Key key,
+    required Widget child,
+    Key? key,
     this.maxHeight,
     this.maxWidth,
     this.minHeight,
@@ -62,7 +61,7 @@ class CssSizing extends SingleChildRenderObjectWidget {
     this.preferredAxis,
     this.preferredHeight,
     this.preferredWidth,
-  })  : assert(child != null),
+  })  : 
         super(child: child, key: key);
 
   @override
@@ -104,7 +103,7 @@ class CssSizing extends SingleChildRenderObjectWidget {
   }
 
   void _debugFillProperty(DiagnosticPropertiesBuilder properties, String name,
-      CssSizingValue value) {
+      CssSizingValue? value) {
     if (value == null) return;
     properties.add(DiagnosticsProperty<CssSizingValue>(name, value));
   }
@@ -124,14 +123,14 @@ class CssSizing extends SingleChildRenderObjectWidget {
 
 class _RenderCssSizing extends RenderProxyBox {
   _RenderCssSizing({
-    RenderBox child,
-    CssSizingValue maxHeight,
-    CssSizingValue maxWidth,
-    CssSizingValue minHeight,
-    CssSizingValue minWidth,
-    Axis preferredAxis,
-    CssSizingValue preferredHeight,
-    CssSizingValue preferredWidth,
+    RenderBox? child,
+    CssSizingValue? maxHeight,
+    CssSizingValue? maxWidth,
+    CssSizingValue? minHeight,
+    CssSizingValue? minWidth,
+    Axis? preferredAxis,
+    CssSizingValue? preferredHeight,
+    CssSizingValue? preferredWidth,
   })  : _maxHeight = maxHeight,
         _maxWidth = maxWidth,
         _minHeight = minHeight,
@@ -141,15 +140,15 @@ class _RenderCssSizing extends RenderProxyBox {
         _preferredWidth = preferredWidth,
         super(child);
 
-  CssSizingValue _maxHeight;
-  CssSizingValue _maxWidth;
-  CssSizingValue _minHeight;
-  CssSizingValue _minWidth;
+  CssSizingValue? _maxHeight;
+  CssSizingValue? _maxWidth;
+  CssSizingValue? _minHeight;
+  CssSizingValue? _minWidth;
   void setConstraints({
-    CssSizingValue maxHeight,
-    CssSizingValue maxWidth,
-    CssSizingValue minHeight,
-    CssSizingValue minWidth,
+    CssSizingValue? maxHeight,
+    CssSizingValue? maxWidth,
+    CssSizingValue? minHeight,
+    CssSizingValue? minWidth,
   }) {
     if (maxHeight == _maxHeight &&
         maxWidth == _maxWidth &&
@@ -164,11 +163,11 @@ class _RenderCssSizing extends RenderProxyBox {
     markNeedsLayout();
   }
 
-  Axis _preferredAxis;
-  CssSizingValue _preferredHeight;
-  CssSizingValue _preferredWidth;
+  Axis? _preferredAxis;
+  CssSizingValue? _preferredHeight;
+  CssSizingValue? _preferredWidth;
   void setPreferredSize(
-      Axis axis, CssSizingValue width, CssSizingValue height) {
+      Axis? axis, CssSizingValue? width, CssSizingValue? height) {
     if (axis == _preferredAxis &&
         height == _preferredHeight &&
         width == _preferredWidth) return;
@@ -218,15 +217,15 @@ class _RenderCssSizing extends RenderProxyBox {
       minWidth: stableChildSize?.width ?? preferredWidth ?? minWidth,
     );
 
-    child.layout(cc, parentUsesSize: true);
-    size = constraints.constrain(child.size);
+    child!.layout(cc, parentUsesSize: true);
+    size = constraints.constrain(child!.size);
   }
 
-  Size _guessChildSize({
-    double maxHeight,
-    double maxWidth,
-    double preferredHeight,
-    double preferredWidth,
+  Size? _guessChildSize({
+    double? maxHeight,
+    double? maxWidth,
+    required double preferredHeight,
+    required double preferredWidth,
   }) {
     final ccHeight = BoxConstraints(
       maxWidth: double.infinity,
@@ -234,8 +233,8 @@ class _RenderCssSizing extends RenderProxyBox {
       minWidth: 0,
       minHeight: preferredHeight,
     );
-    child.layout(ccHeight, parentUsesSize: true);
-    final sizeHeight = child.size;
+    child!.layout(ccHeight, parentUsesSize: true);
+    final sizeHeight = child!.size;
 
     final ccWidth = BoxConstraints(
       maxWidth: preferredWidth,
@@ -243,8 +242,8 @@ class _RenderCssSizing extends RenderProxyBox {
       minWidth: preferredWidth,
       minHeight: 0,
     );
-    child.layout(ccWidth, parentUsesSize: true);
-    final sizeWidth = child.size;
+    child!.layout(ccWidth, parentUsesSize: true);
+    final sizeWidth = child!.size;
 
     final childAspectRatio = sizeWidth.width / sizeWidth.height;
     const epsilon = 0.01;
@@ -254,7 +253,7 @@ class _RenderCssSizing extends RenderProxyBox {
     }
 
     // child appears to have a stable aspect ratio
-    double childWidth, childHeight;
+    double? childWidth, childHeight;
     if (_preferredAxis == Axis.vertical) {
       childHeight = preferredHeight;
       childWidth = childHeight * childAspectRatio;
@@ -263,11 +262,11 @@ class _RenderCssSizing extends RenderProxyBox {
       childHeight = childWidth / childAspectRatio;
     }
 
-    if (childWidth > maxWidth) {
+    if (childWidth > maxWidth!) {
       childWidth = maxWidth;
       childHeight = childWidth / childAspectRatio;
     }
-    if (childHeight > maxHeight) {
+    if (childHeight > maxHeight!) {
       childHeight = maxHeight;
       childWidth = childHeight * childAspectRatio;
     }
@@ -279,7 +278,7 @@ class _RenderCssSizing extends RenderProxyBox {
 /// A [CssSizing] value.
 abstract class CssSizingValue {
   const CssSizingValue._();
-  double clamp(double min, double max);
+  double? clamp(double min, double max);
 
   /// Creates an auto value.
   factory CssSizingValue.auto() = _CssSizingAuto;
@@ -294,7 +293,7 @@ abstract class CssSizingValue {
 class _CssSizingAuto extends CssSizingValue {
   const _CssSizingAuto() : super._();
   @override
-  double clamp(double _, double __) => null;
+  double? clamp(double _, double __) => null;
 
   @override
   int get hashCode => 0;
