@@ -921,7 +921,7 @@ void main() async {
               );
 
               await screenMatchesGolden(tester, testCase.key);
-            }, skip: null);
+            }, skip: false);
           }
         }, skip: Platform.isLinux ? null : 'Linux only');
       },
@@ -935,7 +935,7 @@ void main() async {
 class _Golden extends StatelessWidget {
   final String contents;
 
-  const _Golden(this.contents, {Key key}) : super(key: key);
+  const _Golden(this.contents, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext _) => Scaffold(
@@ -954,10 +954,10 @@ class _Golden extends StatelessWidget {
 }
 
 class _HitTestApp extends StatelessWidget {
-  final String href;
-  final void Function(String) onTapUrl;
+  final String? href;
+  final void Function(String)? onTapUrl;
 
-  const _HitTestApp({this.href, Key key, this.onTapUrl}) : super(key: key);
+  const _HitTestApp({this.href, Key? key, this.onTapUrl}) : super(key: key);
 
   @override
   Widget build(BuildContext _) => MaterialApp(
