@@ -2,15 +2,15 @@ part of '../core_ops.dart';
 
 const kCssPadding = 'padding';
 
-WidgetPlaceholder _paddingInlineAfter(TextStyleBuilder tsb, CssLengthBox box) =>
-    WidgetPlaceholder<CssLengthBox>(box).wrapWith((context, _) =>
-        _paddingInlineSizedBox(box.getValueRight(tsb.build(context))));
+WidgetPlaceholder _paddingInlineAfter(TextStyleBuilder? tsb, CssLengthBox? box) =>
+    WidgetPlaceholder<CssLengthBox?>(box).wrapWith((context, _) =>
+        _paddingInlineSizedBox(box!.getValueRight(tsb!.build(context))));
 
-WidgetPlaceholder _paddingInlineBefore(TextStyleBuilder tsb, CssLengthBox b) =>
-    WidgetPlaceholder<CssLengthBox>(b).wrapWith((context, _) =>
-        _paddingInlineSizedBox(b.getValueLeft(tsb.build(context))));
+WidgetPlaceholder _paddingInlineBefore(TextStyleBuilder? tsb, CssLengthBox? b) =>
+    WidgetPlaceholder<CssLengthBox?>(b).wrapWith((context, _) =>
+        _paddingInlineSizedBox(b!.getValueLeft(tsb!.build(context))));
 
-Widget _paddingInlineSizedBox(double width) =>
+Widget _paddingInlineSizedBox(double? width) =>
     width != null && width > 0 ? SizedBox(width: width) : widget0;
 
 class StylePadding {
@@ -22,20 +22,20 @@ class StylePadding {
 
   BuildOp get buildOp => BuildOp(
         onTree: (meta, tree) {
-          if (meta.willBuildSubtree) return;
+          if (meta.willBuildSubtree!) return;
           final padding = tryParseCssLengthBox(meta, kCssPadding);
           if (padding?.hasLeftOrRight != true) return;
 
           return wrapTree(
             tree,
             append: (p) =>
-                WidgetBit.inline(p, _paddingInlineAfter(p.tsb, padding)),
+                WidgetBit.inline(p, _paddingInlineAfter(p!.tsb, padding)),
             prepend: (p) =>
-                WidgetBit.inline(p, _paddingInlineBefore(p.tsb, padding)),
+                WidgetBit.inline(p, _paddingInlineBefore(p!.tsb, padding)),
           );
         },
         onWidgets: (meta, widgets) {
-          if (widgets?.isNotEmpty != true) return null;
+          if (widgets.isNotEmpty != true) return null;
           final padding = tryParseCssLengthBox(meta, kCssPadding);
           if (padding == null) return null;
 
@@ -49,7 +49,7 @@ class StylePadding {
 
   Widget _build(BuildContext context, BuildMetadata meta, Widget child,
       CssLengthBox padding) {
-    final tsh = meta.tsb().build(context);
+    final tsh = meta.tsb()!.build(context);
     return wf.buildPadding(
       meta,
       child,
